requirements.txt
django==5.0.7
whitenoise==6.6.0
gunicorn==22.0.0
psycopg[binary]==3.2.10
dj-database-url==2.3.0
python-dotenv==1.0.1
django-environ==0.11.2
requests==2.32.3
beautifulsoup4==4.12.3
defusedxml==0.7.1
azure-communication-email==1.0.0
django-tinymce==4.0.0
django-ckeditor==6.7.1
Pillow==10.4.0
django-storages[boto3]==1.14.4
boto3==1.35.36
